refactor(data-ingestion-core): reuse common read-write logic in history
