+++ crates/iota-data-ingestion-core/src/history/mod.rs
@@ -59,6 +59,19 @@
 //! └──────────────────────────────┘
 //! ```
 
+use std::io::{BufWriter, Cursor, Read, Seek, SeekFrom, Write};
+
+use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
+use bytes::Bytes;
+use fastcrypto::hash::{HashFunction, Sha3_256};
+use iota_storage::{
+    SHA3_BYTES,
+    blob::{Blob, BlobEncoding},
+};
+use serde::{Serialize, de::DeserializeOwned};
+
+use crate::{IngestionError, errors::IngestionResult as Result};
+
 pub mod epoch_boundaries;
 pub mod manifest;
 pub mod reader;
@@ -70,3 +83,67 @@ pub const MANIFEST_FILE_MAGIC: u32 = 0x0000FACE;
 pub const MANIFEST_FILENAME: &str = "MANIFEST";
 pub const EPOCH_BOUNDARIES_FILE_MAGIC: u32 = 0x0000E90C;
 pub const EPOCH_BOUNDARIES_FILENAME: &str = "EPOCH_BOUNDARIES";
+
+/// Encodes `value` as a BCS [`Blob`] framed with a 4-byte `magic` prefix and a
+/// trailing SHA3-256 checksum computed over the magic and the blob.
+///
+/// This is the on-disk format shared by the MANIFEST and EPOCH_BOUNDARIES
+/// files.
+pub(crate) fn finalize_magic_blob<T: Serialize>(value: &T, magic: u32) -> Result<Bytes> {
+    let mut buf = BufWriter::new(vec![]);
+    buf.write_u32::<BigEndian>(magic)?;
+    Blob::encode(value, BlobEncoding::Bcs)?.write(&mut buf)?;
+    buf.flush()?;
+    let mut hasher = Sha3_256::default();
+    hasher.update(buf.get_ref());
+    let computed_digest = hasher.finalize().digest;
+    buf.write_all(&computed_digest)?;
+    Ok(Bytes::from(buf.into_inner().map_err(|e| e.into_error())?))
+}
+
+/// Decodes a value written by [`finalize_magic_blob`] and verifies its
+/// integrity.
+///
+/// `filename` names the file in error messages (e.g. `"manifest"`).
+///
+/// # Errors
+///
+/// Fails if the `magic` prefix or the trailing SHA3-256 checksum does not
+/// match.
+pub(crate) fn read_magic_blob<T: DeserializeOwned>(
+    vec: Vec<u8>,
+    magic: u32,
+    filename: &str,
+) -> Result<T> {
+    let file_size = vec.len();
+    let mut reader = Cursor::new(vec);
+
+    // Reads from the beginning of the file and verifies the magic byte.
+    let found = reader.read_u32::<BigEndian>()?;
+    if found != magic {
+        return Err(IngestionError::HistoryRead(format!(
+            "unexpected magic byte in {filename}: {found}",
+        )));
+    }
+
+    // Reads the SHA3 checksum stored at the end of the file.
+    reader.seek(SeekFrom::End(-(SHA3_BYTES as i64)))?;
+    let mut sha3_digest = [0u8; SHA3_BYTES];
+    reader.read_exact(&mut sha3_digest)?;
+
+    // Reads the content and verifies it against the stored checksum.
+    reader.rewind()?;
+    let mut content_buf = vec![0u8; file_size - SHA3_BYTES];
+    reader.read_exact(&mut content_buf)?;
+    let mut hasher = Sha3_256::default();
+    hasher.update(&content_buf);
+    let computed_digest = hasher.finalize().digest;
+    if computed_digest != sha3_digest {
+        return Err(IngestionError::HistoryRead(format!(
+            "{filename} corrupted, computed checksum: {computed_digest:?}, stored checksum: {sha3_digest:?}"
+        )));
+    }
+
+    reader.seek(SeekFrom::Start(MAGIC_BYTES as u64))?;
+    Ok(Blob::read(&mut reader)?.decode()?)
+}
